dts: mt-connect: move pinctrl to clkin_gpt node for now
It's not necessarily wrong to do it this way, but it is more
configurable to have the pinctrl in the clock node, because then you
can just change the clock without touching the pin config. I am still
investigating why it didn't work when pinctrl was on the clock node. I
don't see anything else configuring that pin, but maybe I missed
something, I'm just not sure...