sound: enable revived ak4432
This commit contains the bulk of the changes. I have enabled as much
functionality as I could, but have left some issues open:

* runtime TDM/DIF configuration: There is a table with 16 different
configurations. We need to be able to interpret info from userspace
into the TDM config. We also may need to reinitialize the chip to make
this work. Currently setting this at runtime breaks stuff.
* soft mute: I have heard this cause stuttering and hangs. It actually
could be related to pipewire/dsoundcard underflows. More investigation
required.
* runtime DAPM configuration: We don't use this so I decided to
disable it for now.