MT-132256: revive ak4432 driver from Asahi-Kasei (#34)
* sound: ak4432: whitespace cleanup

* sound: ak4432: use newest ASoC architecture

The new architecture is much more modern. Codecs are separated into
"components".

* dts: mt-connect: add ak4432 dt node and pinmux

* imx-card: add case for ak4432

* sound: enable revived ak4432

This commit contains the bulk of the changes. I have enabled as much
functionality as I could, but have left some issues open:

* runtime TDM/DIF configuration: There is a table with 16 different
configurations. We need to be able to interpret info from userspace
into the TDM config. We also may need to reinitialize the chip to make
this work. Currently setting this at runtime breaks stuff.
* soft mute: I have heard this cause stuttering and hangs. It actually
could be related to pipewire/dsoundcard underflows. More investigation
required.
* runtime DAPM configuration: We don't use this so I decided to
disable it for now.

* ak4432: disable debug mode

I have added tons of debug to this driver. It is very informative for
understanding what exactly happens when we play media from
userspace. There are lots of little interactions and information on
the audio sample that should be handled by the codec

* config: mt-connect: enable ak4432 as module